// apps/backend/src/gamification/constants/xp-events.enum.ts
export enum GamificationEvent {
  ISTQB_COMPLETED = 'ISTQB_COMPLETED',
  ISTQB_PASSED = 'ISTQB_PASSED',
  ISTQB_HIGH_SCORE = 'ISTQB_HIGH_SCORE',
  ALLPAIRS_GENERATED = 'ALLPAIRS_GENERATED',
  ALLPAIRS_LARGE_RESULT = 'ALLPAIRS_LARGE_RESULT',
  DAILY_LOGIN = 'DAILY_LOGIN',
  STREAK_7_DAYS = 'STREAK_7_DAYS',
  ACHIEVEMENT_BONUS = 'ACHIEVEMENT_BONUS',
  THREAD_CREATED = 'THREAD_CREATED',
  POST_CREATED = 'POST_CREATED',
  IDEA_CREATED = 'IDEA_CREATED',
  IDEA_VOTED = 'IDEA_VOTED',
  COMMENT_ADDED = 'COMMENT_ADDED',
  PERFORMANCE_COMPLETED = 'PERFORMANCE_COMPLETED',
  PERFORMANCE_PASSED = 'PERFORMANCE_PASSED',
}

export enum AchievementCriteria {
  ISTQB_COMPLETED_COUNT = 'ISTQB_COMPLETED_COUNT',
  ISTQB_PASSED_COUNT = 'ISTQB_PASSED_COUNT',
  ALLPAIRS_COUNT = 'ALLPAIRS_COUNT',
  ALLPAIRS_LARGE_COUNT = 'ALLPAIRS_LARGE_COUNT',
  DAILY_LOGIN_COUNT = 'DAILY_LOGIN_COUNT',
  STREAK_DAYS = 'STREAK_DAYS',
  USER_LEVEL = 'USER_LEVEL',
}